requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-multipart

